+++ lib/shore/v1/appointment_group.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+require_relative 'client_base'
+
+module Shore
+  module V1
+    class AppointmentGroup < ClientBase
+    end
+  end
+end